requirements.txt
annotated-types==0.7.0
anyio==3.5.0
appdirs==1.4.4
apt-clone==0.2.1
apturl==0.5.2
attrs==24.2.0
Babel==2.8.0
beautifulsoup4==4.10.0
beniget==0.4.1
binwalk==2.3.3
BitVector==3.5.0
black==21.10b0
blinker==1.4
Brlapi==0.8.3
Brotli==1.0.9
certifi==2024.8.30
chardet==4.0.0
charset-normalizer==2.0.6
click==8.0.3
colorama==0.4.4
command-not-found==0.3
commonmark==0.9.1
configobj==5.0.6
cov-core==1.15.0
coverage==6.2
cryptography==3.4.8
cupshelpers==1.0
cycler==0.11.0
dbus-python==1.2.18
decorator==4.4.2
defer==1.0.6
distro==1.7.0
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
exceptiongroup==1.2.2
execnet==1.9.0
eyeD3==0.8.10
filelock==3.6.0
flake8==4.0.1
fonttools==4.29.1
fs==2.4.12
gast==0.5.2
grpcio==1.64.1
grpcio-tools==1.64.1
h11==0.13.0
h2==4.1.0
hpack==4.0.0
html5lib==1.1
httpcore==1.0.5
httplib2==0.20.2
httpx==0.27.0
hyperframe==6.0.1
idna==3.3
ifaddr==0.1.7
IMDbPY==2021.4.18
importlib-metadata==4.6.4
iniconfig==1.1.1
isort==5.6.4
jeepney==0.7.1
Jinja2==3.0.3
jsonschema==3.2.0
keyring==23.5.0
kiwisolver==1.3.2
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
livereload==2.6.3
louis==3.20.0
lxml==4.8.0
lz4==3.1.3+dfsg
macaroonbakery==1.3.1
Mako==1.1.3
Markdown==3.3.6
MarkupSafe==2.0.1
matplotlib==3.5.1
mccabe==0.6.1
mkdocs==1.1.2
more-itertools==8.10.0
mpmath==0.0.0
mypy==0.942
mypy-extensions==0.4.3
nala==0.11.1
nemo-emblems==5.6.0
netaddr==0.8.0
netifaces==0.11.0
nose2==0.9.2
numpy==1.21.5
oauthlib==3.2.0
onboard==1.4.1
openai==1.35.10
outcome==1.3.0.post0
packaging==21.3
PAM==0.4.2
pathspec==0.9.0
pexpect==4.8.0
Pillow==9.0.1
platformdirs==2.5.1
pluggy==0.13.0
ply==3.11
portalocker==2.10.0
protobuf==3.20.1
psutil==5.9.0
ptyprocess==0.7.0
py==1.10.0
pycairo==1.20.1
pycodestyle==2.8.0
pycups==2.0.1
pycurl==7.44.1
pydantic==2.8.2
pydantic_core==2.20.1
pyelftools==0.27
pyflakes==2.4.0
Pygments==2.11.2
PyGObject==3.42.1
PyICU==2.8.1
pyinotify==0.9.6
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
PyOpenGL==3.1.5
pyparsing==2.4.7
pyparted==3.11.7
PyQt5==5.15.6
PyQt5-sip==12.9.1
pyqtgraph==0.12.4
pyRFC3339==1.1
pyrsistent==0.18.1
PySocks==1.7.1
pytest==6.2.5
pytest-cov==3.0.0
pytest-forked==1.4.0
pytest-sugar==0.9.4
pytest-xdist==2.5.0
python-apt==2.4.0+ubuntu2
python-dateutil==2.8.1
python-debian==0.1.43+ubuntu1.1
python-dotenv==0.19.2
python-gnupg==0.4.8
python-magic==0.4.24
python-xlib==0.29
pythran==0.10.0
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
qdrant-client==1.10.1
regex==2024.7.24
reportlab==3.6.8
requests==2.32.3
requests-file==1.5.1
rfc3986==1.5.0
rich==11.2.0
scipy==1.8.0
SecretStorage==3.3.1
selenium==4.25.0
setproctitle==1.2.2
shellingham==1.4.0
six==1.16.0
sniffio==1.3.1
socksio==1.0.0
sortedcontainers==2.4.0
soupsieve==2.3.1
sympy==1.9
systemd-python==234
termcolor==1.1.0
texttable==1.6.4
tiktoken==0.7.0
tinycss2==1.1.1
tldextract==3.1.2
toml==0.10.2
tomli==1.2.2
tornado==6.1
tqdm==4.66.4
trash-cli==0.17.1.14
trio==0.26.2
trio-websocket==0.11.1
typed-ast==1.4.3
typer==0.4.0
typing_extensions==4.12.2
ubuntu-drivers-common==0.0.0
ufoLib2==0.13.1
ufw==0.36.1
undetected-chromedriver==3.5.5
unicodedata2==14.0.0
Unidecode==1.3.3
urllib3==2.2.2
vboxapi==1.0
wadllib==1.3.6
webencodings==0.5.1
websocket-client==1.8.0
websockets==13.1
wsproto==1.2.0
xdg==5
xkit==0.0.0
xlrd==1.2.0
youtube-dl==2021.12.17
zipp==1.0.0
